Created UsersInvitationApp
This is an app, which has a list of users. In front of every user you can see a "plus" sign. If you click this sign, the user will be added to list of users, you can send an invitation by pressing a button below, and the sign will change to "minus' sign.